requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
requests==2.31.0
flask-wtf==1.2.1
flask-talisman==1.1.0
flask-limiter==3.5.0 